// frontend/src/app/core/model/model.d.ts
export interface INotepad {

    /** Notepad id */
    id?: (number|null);

    /** Notepad isDeleted */
    isDeleted?: (boolean|null);

    /** Notepad noteAmount */
    noteAmount?: (number|null);

    /** Notepad created */
    created: number;

    /** Notepad updated */
    updated: number;

    /** Notepad name */
    name: string;
}

/** Represents a Notepad. */
export class Notepad implements INotepad {

    /**
     * Constructs a new Notepad.
     * @param [properties] Properties to set
     */
    constructor(properties?: INotepad);

    /** Notepad id. */
    public id: number;

    /** Notepad isDeleted. */
    public isDeleted: boolean;

    /** Notepad noteAmount. */
    public noteAmount: number;

    /** Notepad created. */
    public created: number;

    /** Notepad updated. */
    public updated: number;

    /** Notepad name. */
    public name: string;

    /**
     * Creates a new Notepad instance using the specified properties.
     * @param [properties] Properties to set
     * @returns Notepad instance
     */
    public static create(properties?: INotepad): Notepad;

    /**
     * Encodes the specified Notepad message. Does not implicitly {@link Notepad.verify|verify} messages.
     * @param message Notepad message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encode(message: INotepad, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Decodes a Notepad message from the specified reader or buffer.
     * @param reader Reader or buffer to decode from
     * @param [length] Message length if known beforehand
     * @returns Notepad
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): Notepad;

    /**
     * Creates a Notepad message from a plain object. Also converts values to their respective internal types.
     * @param object Plain object
     * @returns Notepad
     */
    public static fromObject(object: { [k: string]: any }): Notepad;

    /**
     * Creates a plain object from a Notepad message. Also converts values to other types if specified.
     * @param message Notepad
     * @param [options] Conversion options
     * @returns Plain object
     */
    public static toObject(message: Notepad, options?: $protobuf.IConversionOptions): { [k: string]: any };

    /**
     * Converts this Notepad to JSON.
     * @returns JSON object
     */
    public toJSON(): { [k: string]: any };
}

/** Properties of a Note. */
export interface INote {

    /** Note id */
    id?: (number|null);

    /** Note isDeleted */
    isDeleted?: (boolean|null);

    /** Note notepadId */
    notepadId: number;

    /** Note created */
    created: number;

    /** Note updated */
    updated: number;

    /** Note name */
    name: string;

    /** Note text */
    text?: (string|null);

    /** Note isPinned */
    isPinned?: (boolean|null);

    /** Note isArchived */
    isArchived?: (boolean|null);

    /** Note section */
    section?: (string|null);
}

/** Represents a Note. */
export class Note implements INote {

    /**
     * Constructs a new Note.
     * @param [properties] Properties to set
     */
    constructor(properties?: INote);

    /** Note id. */
    public id: number;

    /** Note isDeleted. */
    public isDeleted: boolean;

    /** Note notepadId. */
    public notepadId: number;

    /** Note created. */
    public created: number;

    /** Note updated. */
    public updated: number;

    /** Note name. */
    public name: string;

    /** Note text. */
    public text: string;

    /** Note isPinned. */
    public isPinned: boolean;

    /** Note isArchived. */
    public isArchived: boolean;

    /** Note section. */
    public section: string;

    /**
     * Creates a new Note instance using the specified properties.
     * @param [properties] Properties to set
     * @returns Note instance
     */
    public static create(properties?: INote): Note;

    /**
     * Encodes the specified Note message. Does not implicitly {@link Note.verify|verify} messages.
     * @param message Note message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encode(message: INote, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Decodes a Note message from the specified reader or buffer.
     * @param reader Reader or buffer to decode from
     * @param [length] Message length if known beforehand
     * @returns Note
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): Note;

    /**
     * Creates a Note message from a plain object. Also converts values to their respective internal types.
     * @param object Plain object
     * @returns Note
     */
    public static fromObject(object: { [k: string]: any }): Note;

    /**
     * Creates a plain object from a Note message. Also converts values to other types if specified.
     * @param message Note
     * @param [options] Conversion options
     * @returns Plain object
     */
    public static toObject(message: Note, options?: $protobuf.IConversionOptions): { [k: string]: any };

    /**
     * Converts this Note to JSON.
     * @returns JSON object
     */
    public toJSON(): { [k: string]: any };
}
// ...
export interface IFile {

    /** File contentType */
    contentType?: (string|null);

    /** File extension */
    extension?: (string|null);

    /** File data */
    data?: (string|null);
}

/** Represents a File. */
export class File implements IFile {

    /**
     * Constructs a new File.
     * @param [properties] Properties to set
     */
    constructor(properties?: IFile);

    /** File contentType. */
    public contentType: string;

    /** File extension. */
    public extension: string;

    /** File data. */
    public data: string;

    /**
     * Creates a new File instance using the specified properties.
     * @param [properties] Properties to set
     * @returns File instance
     */
    public static create(properties?: IFile): File;

    /**
     * Encodes the specified File message. Does not implicitly {@link File.verify|verify} messages.
     * @param message File message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encode(message: IFile, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Decodes a File message from the specified reader or buffer.
     * @param reader Reader or buffer to decode from
     * @param [length] Message length if known beforehand
     * @returns File
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): File;

    /**
     * Creates a File message from a plain object. Also converts values to their respective internal types.
     * @param object Plain object
     * @returns File
     */
    public static fromObject(object: { [k: string]: any }): File;

    /**
     * Creates a plain object from a File message. Also converts values to other types if specified.
     * @param message File
     * @param [options] Conversion options
     * @returns Plain object
     */
    public static toObject(message: File, options?: $protobuf.IConversionOptions): { [k: string]: any };

    /**
     * Converts this File to JSON.
     * @returns JSON object
     */
    public toJSON(): { [k: string]: any };
}
// ...
export enum AppMode {
    FINANCE = 0,
    NOTES = 1
}